Changing directory tree
Put all components of Teste folder a level up

Commit:Mod

TaskNumber:#49
