+++ Monitores/SC_outmod.cpp
