--- Teste/Monitores/SC_common.h
@@ -1,14 +1,0 @@
-#ifndef SC_COMMON_MOD
-#define SC_COMMON_MOD
-
-#define TAM_FLIT 16
-#define NUM_ROT_X 2
-#define NUM_ROT_Y 2
-#define NUM_EP NUM_ROT_X*NUM_ROT_Y
-
-#include "systemc.h"
-#include <stdio.h>
-#include <string.h>
-#include <sys/timeb.h>
-
-#endif// SC_COMMON_MOD
